Database merge

+++ src/main/java/cz/muni/fi/pv168/seminar01/beta/data/storage/dao/VehicleDao.java
@@ -38,7 +38,7 @@ public class VehicleDao implements DataAccessObject<VehicleEntity> {
             statement.setString(2, entity.brand());
             statement.setString(3, entity.type());
             statement.setInt(4, entity.capacity());
-            statement.setFloat(5, entity.consumption());
+            statement.setDouble(5, entity.consumption());
             statement.setString(6, entity.fuelType().name().toLowerCase());
             statement.executeUpdate();
 
@@ -199,7 +199,7 @@ public class VehicleDao implements DataAccessObject<VehicleEntity> {
                 resultSet.getString("brand"),
                 resultSet.getString("type"),
                 resultSet.getInt("capacity"),
-                resultSet.getFloat("consumption"),
+                resultSet.getDouble("consumption"),
                 FuelType.valueOf(resultSet.getString("fuelType").toUpperCase())
         );
     }
